PersistentQueue updated

+++ util/src/main/java/de/uniluebeck/itm/tr/util/FileUtils.java
@@ -68,7 +68,8 @@ public class FileUtils {
 		}
 	}
 
-    public static void deleteDirectory(File dir) {
+    public static void deleteDirectory(File dir) throws SecurityException {
+        if (!dir.exists()) return;
         for (String filename : dir.getAbsoluteFile().list()){
             new File(dir.getAbsolutePath(), filename).delete();
         }
